+++ src/hooks/useValidation.ts
@@ -138,13 +138,11 @@ const useValidation = (value: string, validations:TValidations) => {
 
         case 'isCardDateMonthInvalid':
           if (value.trim().length === 1 && !/^([1-9])$/.test(value.trim())) {
-            console.log('here');
             setIsCardDateMonthInvalid(true);
             break;
           } else setIsCardDateMonthInvalid(false);
 
           if (value.trim().length > 1 && !/^(1[012]|0?[1-9])$/.test(value)) {
-            console.log('there');
             setIsCardDateMonthInvalid(true);
           } else setIsCardDateMonthInvalid(false);
           break;
